requirements.txt
langchain
openai
python-dotenv
pinecone
tiktoken
pandas